Add Card component and related hooks
This commit includes the following changes:

- Created a new Card component with associated subcomponents (CardDescription and CardTitle).
- Added a new hooks folder with custom hooks (useEffectOnUpdate and useToggle).
- Updated App.jsx, main.jsx, and styles.css to incorporate the new Card component.

Example of use of card component:
    <Card bgIcon="red" width="400px" onClick={() => console.log("Opened/closed")}>
      <Card.Title>Easy Deployment</Card.Title>
      <Card.Description>Ac tincidunt sapien vehicula erat auctor pellentesque rhoncus. Et magna sit morbi lobortis.</Card.Description>
    </Card>

The Card component accepts the icon, bgIcon, width and onClick props. All four of them are not mandatory, a default will be used.
Also allows children props, and the CardTitle subcomponent os mandatory, but the CardDescription subcomponent is not mandatory.
If CardDescription is not indicated, the Card will display only the title.
The icon can be clicked and it will expand the description if there's any. Aside the onClick prop of the card component will run the function we provide every time the component updates.